rename single_text_select_question_input.svelte
Renames `singletextselect_question_input.svelte` to `single_text_select_question_input.svelte`.
