+++ src/lib/components/single_text_select_question_input.svelte
